Fix new upload photo

+++ Calendary.Ng/src/app/components/calendar-images/calendar-images.component.ts
@@ -1,5 +1,5 @@
 import { CommonModule } from '@angular/common';
-import { Component, EventEmitter, Input, OnChanges, OnInit, Output, SimpleChanges } from '@angular/core';
+import { Component, ElementRef, EventEmitter, Input, OnChanges, OnInit, Output, SimpleChanges, ViewChild } from '@angular/core';
 import { FileUploader, FileUploadModule } from 'ng2-file-upload';
 import { ImageService } from '../../../services/image.service';
 import { Image } from '../../../models/image';
@@ -14,18 +14,21 @@ import { ImageItemComponent } from '../image-item/image-item.component';
   styleUrl: './calendar-images.component.scss'
 })
 export class CalendarImagesComponent implements OnChanges, OnInit {
+
+
   @Input() calendarId!: number;
+  @Output() uploadCompleted = new EventEmitter<string>();
+  @Output() generatedCompleted = new EventEmitter<boolean>();
+  
+  @ViewChild('fileInput') fileInput!: ElementRef;
 
   items = Array.from({length: 12}, (_, i) => i + 1);
   monthes = ['Січень', 'Лютий', 'Березень', 'Квітень', 'Травень', 'Червень', 'Липень', 'Серпень', 'Вересень', 'Жовтень', 'Листопад', 'Грудень'];
   images: Image[] = [];
 
   public uploader!: FileUploader;
+  selectedMonth: number = 0;
   
-  @Output() uploadCompleted = new EventEmitter<string>();
-  @Output() generatedCompleted = new EventEmitter<boolean>();
-
-
   constructor(private imageService : ImageService, 
     private calendarService: CalendarService
   ) 
@@ -40,7 +43,8 @@ export class CalendarImagesComponent implements OnChanges, OnInit {
 
 
   ngOnChanges(changes: SimpleChanges): void {
-    if (changes['calendarId'] && !changes['calendarId'].firstChange) {
+    if (changes['calendarId'] && !changes['calendarId'].firstChange) 
+    {
       this.initializeUploader();
     }
   }
@@ -55,6 +59,10 @@ export class CalendarImagesComponent implements OnChanges, OnInit {
         maxFileSize: 5 * 1024 * 1024 // Максимальний розмір 5MB
       });
 
+      this.uploader.onBeforeUploadItem = (item) => {
+        item.url = `/api/image/upload/${this.calendarId}?month=${this.selectedMonth}`;
+      };
+
       this.uploader.onCompleteItem = (item, response, status, headers) => {
         this.loadImages();
       };
@@ -65,8 +73,8 @@ export class CalendarImagesComponent implements OnChanges, OnInit {
     }
   }
 
-  getImage(index : number): Image | undefined {
-    return this.images.find(i => i.order === index);
+  getImage(monthNumber : number): Image | undefined {
+    return this.images.find(i => i.monthNumber === monthNumber);
   }
 
   loadImages(): void {
@@ -81,8 +89,17 @@ export class CalendarImagesComponent implements OnChanges, OnInit {
     );
   }
 
-  deleteImage(index: number): void {
-    const image = this.images[index];
+  onStartUpload(monthNumber: number): void 
+  {
+    this.selectedMonth = monthNumber;
+    this.fileInput.nativeElement.click();
+  }
+
+  deleteImage(monthNumber: number): void {
+    const image = this.images.find(i => i.monthNumber === monthNumber);
+    if (!image) {
+      return;
+    }
     this.imageService.deleteImage(image.id).subscribe(() => {
       this.loadImages();
     });
